Added exception

+++ src/library/books.controller.ts
@@ -1,4 +1,6 @@
-import { Controller, Post, Body, Get, Put, Param, Delete } from '@nestjs/common';
+import { Controller, Post, Body, Get, Put, Param, Delete, UsePipes, HttpException, UseFilters } from '@nestjs/common';
+import {ValidationPipe} from "./validation/validation.pipe";
+import {LoginDto} from "./dto/login.dto";
 import { BookDocument } from './schemas/books.schema';
 import { BooksService } from './books.service';
 import { HydratedDocument, QueryWithHelpers } from 'mongoose';
@@ -6,7 +8,9 @@ import { HydratedDocument, QueryWithHelpers } from 'mongoose';
 import { IParamId } from './interfaces/param-id';
 import { CreateBookDto } from './interfaces/dto/create-book';
 import { UpdateBookDto } from './interfaces/dto/update-book';
+import { HttpExceptionFilter } from '../http.exception.filter';
 
+@UseFilters(HttpExceptionFilter)
 @Controller('books')
 export class BooksController {
     constructor(private readonly bookService: BooksService) {}
@@ -31,6 +35,21 @@ export class BooksController {
 
     @Delete(':id')
     public delete(@Param() { id }: IParamId): QueryWithHelpers<HydratedDocument<BookDocument, {}, {}> | null, HydratedDocument<BookDocument, {}, {}>, {}, BookDocument> {
-        return this.bookService.delete(id);
+        if(id == 'wrong-test') {
+            throw new HttpException('Oops', 401);
+        }
+        else {
+            return this.bookService.delete(id);
+        }
+
+    }
+
+    @UsePipes(new ValidationPipe())
+    @Post('/login')
+    login(@Body() body: LoginDto) {
+      return body;
     }
 }
+
+
+
